@@ -20,9 +20,9 @@
         "semver": "^7.6.0",
         "sort-package-json": "^2.8.0",
         "tsx": "^4.19.0",
-        "turbo": "^1.12.3",
+        "turbo": "^2.1.2",
         "type-fest": "^4.10.3",
-        "typescript": ">=3.0.0",
+        "typescript": "^5.6.2",
         "util": "^0.12.3"
     }
 }
