@@ -106,7 +106,7 @@
     "strip-indent": "^4.0.0",
     "tailwindcss": "^3.4.3",
     "tsx": "^4.19.0",
-    "typescript": "^5.2.2",
+    "typescript": "^5.6.2",
     "typescript-eslint": "^7.1.0",
     "vite": "^5.0.0",
     "vitest": "^1.2.2"
