@@ -33,28 +33,32 @@ export function errorFromUnknown(err: unknown) {
   return err instanceof Error ? err : new Error(`Non error thrown: ${String(err)}`, {cause: err})
 }
 
-export class QueryErrorCause {
-  constructor(
-    readonly query: Pick<SQLQuery<unknown>, 'name'> & Partial<SQLQuery<unknown>>,
-    readonly error: Error | undefined,
-    readonly result: {rows: unknown[]} | undefined,
-  ) {}
-}
-
-export class QueryError extends Error {
-  cause!: {
-    name: string
-    message: string
+export namespace QueryError {
+  export type Params = {
     query: Pick<SQLQuery<unknown>, 'name'> & Partial<SQLQuery<unknown>>
-    error?: Error
     result?: {rows: unknown[]}
+    cause?: Error
+  }
+}
+export class QueryError extends Error {
+  query: QueryError.Params['query']
+  result?: QueryError.Params['result']
+
+  constructor(message: string, {query, result, cause}: QueryError.Params) {
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    const suffix = cause && (cause as any).code in pgErrorCodes ? ` (${pgErrorCodes[(cause as any).code]})` : ''
+    super(`[${query.name}]: ${message}${suffix}`, cause ? {cause} : undefined)
+    this.query = query
+    this.result = result
   }
 
-  constructor(message: string, cause: Omit<QueryError['cause'], 'name' | 'message'>) {
-    super(`[Query ${cause.query.name}]: ${message || cause?.error?.message || cause?.error?.constructor?.name}`, {
-      cause,
-    })
-    this.cause = {...cause, message: cause?.error?.message || '', name: 'QueryErrorCause'}
+  toJSON() {
+    return {
+      message: this.message,
+      query: this.query,
+      result: this.result,
+      cause: this.cause,
+    }
   }
 }
 
