@@ -1,5 +1,5 @@
 export {Changes} from './changes'
-export {run, type Flags} from './command'
+export {run, type MigraOptions} from './command'
 export {Migration} from './migra'
 export {Statements, UnsafeMigrationException} from './statements'
 
