@@ -1,3 +1,5 @@
+import {Client} from '@pgkit/client'
+import {type MigratorConstructorParams} from '@pgkit/migrator/dist/types'
 import {type Options as TypegenOptions} from '@pgkit/typegen'
 import * as fs from 'fs'
 import * as path from 'path'
@@ -6,20 +8,22 @@ export type Config = {
   client: {
     connectionString: string
   }
-  typegen?: Partial<TypegenOptions>
-  migrator?: {
-    connectionString?: string
-    /** @default '${cwd}/migrations' */
-    migrationTableName?: string
-    /** @default 'migrations' */
-    migrationsPath?: string
-  }
+  typegen?:
+    | Partial<TypegenOptions>
+    | ((params: {client: Client; defaults: typeof import('@pgkit/typegen').defaults}) => Partial<TypegenOptions>)
+  migrator?:
+    | Omit<MigratorConstructorParams, 'client'>
+    | ((params: {client: Client}) => Omit<MigratorConstructorParams, 'client'>)
+}
+
+export type ResolvedConfig = {
+  [K in keyof Config]: Exclude<Config[K], (...args: never) => unknown>
 }
 
 export const defineConfig = (config: Config) => config
 
 export const loadConfig = async (): Promise<Config> => {
-  const importx = await import('importx')
+  const importx = await import('importx') // todo: consider c12 instead
   const configLocations = ['pgkit.config.ts', 'pgkit.config.js']
   let config: Config | undefined
   let cwd = process.cwd()
@@ -30,7 +34,7 @@ export const loadConfig = async (): Promise<Config> => {
         config = await importx.import(filepath, {
           parentURL: new URL(`file://${cwd}`),
         })
-        if ((config as {default?: Config}).default) {
+        while ((config as {default?: Config}).default) {
           config = (config as {default?: Config}).default
         }
         if (!config) {
